pyproject.toml
[project]
name = "crewai-agents"
version = "0.1.0"
description = "CrewAI agents with AWS Bedrock Nova Micro"
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "crewai>=1.9.3",
    "langchain-aws>=1.2.5",
    "boto3>=1.42.50",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.10",
    "langchain-core>=1.2.13",
]

[dependency-groups]
dev = [
    # Add development dependencies here if needed
]
